All interface variables unsigned

+++ src/mock_align.rs
@@ -323,7 +323,7 @@ fn test_align(reference: &[u8],
             query_ends: query_ends,
             matching: matching,
             indel_cost: INDEL_COST as usize,
-            min_overlap: min_overlap as isize,
+            min_overlap: min_overlap,
         };
     let mut new_aligner = align::Aligner::new(&new_aligner_conf, reference).unwrap();
     let new_actual = new_aligner.locate(query);
@@ -335,7 +335,7 @@ fn test_align(reference: &[u8],
                     && refstop == new_actual_location.refstop() as isize
                     && querystart == new_actual_location.querystart() as isize
                     && querystop == new_actual_location.querystop() as isize
-                    && matches == new_actual_location.matches()
+                    && matches == new_actual_location.matches() as isize
                     && errors == new_actual_location.errors() as isize
             } else {
                 false
